refactored Endpoint router
