--- app/routers/EndpointRouters.js
@@ -1,16 +1,0 @@
-'use strict';
-
-let cors = require ('cors');
-
-const corsOptions = {
-    methods: ['GET', 'POST', 'PUT', 'DELETE', 'SEARCH']
-};
-
-module.exports = {
-
-    '/v1': {
-        use: [
-            cors (corsOptions),
-        ]
-    }
-};
